backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
google-generativeai==0.8.3
slowapi==0.1.9
